Crate hopr_crypto_packet

core-packet

This crate contains the main packet processing functionality for the HOPR protocol. It implements the following important protocol building blocks:

SPHINX packet format (module packet)

Proof of Relay (module por)

Finally, it also implements a utility function which is used to validate tickets (module validation). The ticket validation functionality is dependent on chain-db.

The currently used implementation is selected using the CurrentSphinxSuite type in the packet module.

The implementation can be easily extended for different elliptic curves (or even arithmetic multiplicative groups). In particular, as soon as there’s way to represent Ed448 PeerIDs, it would be easy to create e.g. X448Suite.

Modules

chain

Implements the overlay packet intermediary object.

errors

Enumerates all errors in this crate.

packet

Implements SPHINX packet format.

por

Implements the Proof of Relay.

validation

Implements ticket validation logic.

Type Aliases

CurrentSphinxSuite

Currently used public key cipher suite for Sphinx.