hopr_crypto_packet::packet

Struct PartialHoprPacket

pub struct PartialHoprPacket {
    partial_packet: PartialPacket<HoprSphinxSuite, HoprSphinxHeaderSpec>,
    surbs: Vec<HoprSurb>,
    openers: Vec<HoprReplyOpener>,
    ticket: Ticket,
    next_hop: OffchainPublicKey,
    ack_challenge: HalfKeyChallenge,
}

Represents an outgoing packet that has been only partially instantiated.

It contains PartialPacket, required Proof-of-Relay fields, and the [Ticket], but it does not contain the payload.

This can be used to pre-compute packets for certain destinations, and convert them to full packets once the payload is known.

Fields
partial_packet: PartialPacket<HoprSphinxSuite, HoprSphinxHeaderSpec>surbs: Vec<HoprSurb>openers: Vec<HoprReplyOpener>ticket: Ticketnext_hop: OffchainPublicKeyack_challenge: HalfKeyChallenge
Implementations

impl PartialHoprPacket

pub fn new<M: ProtocolKeyIdMapper<HoprSphinxSuite, HoprSphinxHeaderSpec>, P: NonEmptyPath<OffchainPublicKey> + Send>( pseudonym: &HoprPseudonym, routing: PacketRouting<P>, chain_keypair: &ChainKeypair, ticket: TicketBuilder, mapper: &M, domain_separator: &Hash, ) -> Result<Self>

Instantiates a new partial HOPR packet.

Arguments

pseudonym our pseudonym as packet sender.

routing routing to the destination.

chain_keypair private key of the local node.

ticket ticket builder for the first hop on the path.

mapper of the public key identifiers.

domain_separator channels contract domain separator.

pub fn into_hopr_packet<S: Into<PacketSignals>>( self, msg: &[u8], signals: S, ) -> Result<(HoprPacket, Vec<HoprReplyOpener>)>

Turns this partial HOPR packet into a full Outgoing HoprPacket by attaching the given payload msg and optional packet signals for the recipient.

No signals are equivalent to 0.