Function hopr_crypto_packet::por::pre_verify

pub fn pre_verify(
    secret: &SharedSecret,
    por_bytes: &[u8],
    challenge: &EthereumChallenge
) -> Result<ProofOfRelayOutput>

Verifies that an incoming packet contains all values that are necessary to reconstruct the response to redeem the incentive for relaying the packet.

Arguments

secret shared secret with the creator of the packet

por_bytes serialized ProofOfRelayString as included within the packet

challenge ticket challenge of the incoming ticket