hopr_crypto_packet::sphinx

Module derivation

Contains simple key derivation functions for different purposes

Constants

HASH_KEY_PACKET_TAG 🔒

Functions

create_kdf_instance 🔒

derive_packet_tag

Derives the packet tag used during packet construction by expanding the given secret.

generate_key 🔒

Internal convenience function to generate key and IV from the given secret, that is cryptographically strong.

generate_key_iv 🔒

Internal convenience function to generate key and IV from the given secret, that is cryptographically strong.