hopr_crypto_packet::sphinx

Module packet

Contains the main implementation of a SPHINX packet.

Structs

MetaPacket

An encrypted packet with a payload of size P. The final packet size is given by MetaPacket::SIZE.

PaddedPayload

Holds data that are padded up to P + 1.

PartialPacket

Represents a packet that is only partially instantiated, that is - it contains only the routing information and the Alpha value.

SimpleBiMapper

Basic implementation of the [KeyIdMapping] trait for a simple bi-map.

Enums

ForwardedMetaPacket

Represent a MetaPacket with one layer of encryption removed, exposing the details about the next hop.

MetaPacketRouting

Describes how a MetaPacket should be routed to the destination.

Traits

ProtocolKeyIdMapper

Protocol instantiation specific implementation of the [KeyIdMapping]