hopr_crypto_packet::sphinx

Module routing

Implementation of the SPHINX header format

Structs

HeaderPrefix 🔒

Sphinx header byte prefix

RoutingInfo

Carries routing information for the mixnet packet.

Enums

ForwardedHeader

Enum carry information about the packet based on whether it is destined for the current node (FinalNode) or if the packet is supposed to be only relayed (RelayNode).

Constants

HASH_KEY_PRG 🔒

HASH_KEY_TAG 🔒

SPHINX_HEADER_VERSION 🔒

Current version of the header

Traits

SphinxHeaderSpec

Contains the necessary size and type specifications for the Sphinx packet header.

Functions

forward_header

Applies the forward transformation to the header. If the packet is destined for this node, it returns the additional data for the final destination (ForwardedHeader::Final). Otherwise, it returns the transformed header, the next authentication tag, the public key of the next node, and the additional data for the relayer (ForwardedHeader::Relayed).