Crate hopr_crypto_packet

core-packet

This crate contains the main packet processing functionality for the HOPR protocol. It implements the following important protocol building blocks:

HOPR specific instantiation of the SPHINX packet format

Proof of Relay

Finally, it also implements a utility function which is used to validate tickets (module validation).

The currently used implementation is selected using the HoprSphinxSuite type in the packet module.

The implementation can be easily extended for different elliptic curves (or even arithmetic multiplicative groups). In particular, as soon as there is a way to represent Ed448 PeerIDs, it would be straightforward to create e.g. X448Suite.

This crate implements RFC-0003.

Modules

errors

Lists all errors in this crate.

Structs

HoprSphinxHeaderSpec

Current Sphinx header specification for the HOPR protocol.

ReplyOpener

Entry stored locally by the [SURB] creator to allow decryption of received responses.

Traits

KeyIdMapper

Trait that defines 1:1 mapper between key identifiers and the actual public keys.

Type Aliases

HoprKeyIdent

Type alias for 32-bit HOPR Offchain Public Key Identifier.

HoprReplyOpener

Type alias for identifiable ReplyOpener.

HoprSphinxSuite

Currently used public key cipher suite for Sphinx.

HoprSurb

Single Use Reply Block representation for HOPR protocol.